requirements.txt
newsapi-python
arxiv